+++ applications/calculations/client/js/calculationsClient.js
@@ -8,6 +8,7 @@
  */
 
 (function () {
+  "use strict";
 
   // ------------------------------------------
   // Define an Echo application class.
@@ -300,7 +301,6 @@
   // Initial UI setup.
   jQuery(document).ready(function () {
     app.uiSetup();
-    Thywill.serverInterface.registerApplication(app);
   });
 
 })();