+++ applications/display/client/js/displayClient.js
@@ -10,6 +10,7 @@
  */
 
 (function () {
+  "use strict";
 
   // ------------------------------------------
   // Define a Display application class.
@@ -246,7 +247,6 @@
   // Initial UI setup.
   jQuery(document).ready(function () {
     app.uiSetup();
-    Thywill.serverInterface.registerApplication(app);
   });
 
 })();