+++ applications/draw/client/js/drawClient.js
@@ -12,6 +12,7 @@
  */
 
 (function () {
+  "use strict";
 
   // ------------------------------------------
   // Define a Draw application class.
@@ -271,7 +272,6 @@
   // Initial UI setup.
   jQuery(document).ready(function () {
     app.uiSetup();
-    Thywill.serverInterface.registerApplication(app);
   });
 
 })();