+++ applications/echo/client/js/echoClient.js
@@ -9,6 +9,7 @@
  */
 
 (function () {
+  "use strict";
 
   // ------------------------------------------
   // Define an Echo application class.
@@ -169,7 +170,6 @@
   // Initial UI setup.
   jQuery(document).ready(function () {
     app.uiSetup();
-    Thywill.serverInterface.registerApplication(app);
   });
 
 })();