+++ core/client/socketIoClientInterface/thywillLoadLast.js
@@ -1,8 +1,0 @@
-/*global
-  Thywill: false
-*/
-/**
- * @fileOverview
- * Contains the final Javascript loaded on the bootstrap page.
- */
-Thywill.serverInterface.setupConnection();